bugfix density charge update
Before, the charge integral could only be updated once.

TODO: Right now, this performs the integration every time the charge is
referenced. The more efficient way would be to reset the self._charge
attribute to None whenever the density is update, making the check in
the previous implementation return True. However, the density class
currently doesn't have an update method, making it difficult to automate
this.